feat: Adds support for recipe deletes

+++ src/meal_planner/services/base.py
@@ -182,5 +182,6 @@ class CRUDBase(
 
         """
         record = db.get(self.model, row_id)
-        db.delete(record)
-        db.commit()
+        if record:
+            db.delete(record)
+            db.commit()
